perf(gamefont): Cache one byte of coverage per glyph pixel instead of a full texel (#3288)
Store_GDI_Char composed each cached pixel as (v ? 0x0FFF : 0) | ((v >> 4) << 12)
from the 8 bit GDI coverage v, so of the 16 stored bits only the alpha nibble and
whether the coverage was non zero carried any information.

Store v itself and let Blit_Char rebuild the texel. The reconstruction is exact,
including the transparent white pixels that a coverage below one alpha step
produces, so rendering is unchanged bit for bit. Blit_Char runs when a sentence is
rebuilt rather than per frame, so the added work per pixel does not matter.

A glyph block now holds the same number of glyphs in half the bytes, which is what
keeps large point sizes affordable.

+++ Core/Libraries/Source/WWVegas/WW3D2/render2dsentence.h
@@ -57,7 +57,7 @@ class FontCharsClassCharDataStruct
 public:
 	WCHAR				Value;
 	short				Width;
-	uint16 *		Buffer;
+	uint8 *			Buffer;
 };
 
 enum { CHAR_BUFFER_LEN		= 32768 };
@@ -66,13 +66,13 @@ class FontCharsBuffer
 {
 public:
 	FontCharsBuffer() : Length( 0 ), Buffer( nullptr ) {}
-	FontCharsBuffer( int length, uint16 *buffer ) : Length( length ), Buffer( buffer ) {}
+	FontCharsBuffer( int length, uint8 *buffer ) : Length( length ), Buffer( buffer ) {}
 
 	bool operator== (const FontCharsBuffer &src) const { return Length == src.Length && Buffer == src.Buffer; }
 	bool operator!= (const FontCharsBuffer &src) const { return !(*this == src); }
 
 	int				Length;
-	uint16 *		Buffer;
+	uint8 *			Buffer;
 };
